fix: resolve fast-refresh ESLint warnings by separating non-component exports
- Extract useAuth hook and AuthContext to auth/useAuth.js (was co-located with AuthGate component)
- Extract buildAttributeEntries to utils/attributeEntries.js (was in EntityDetailLayout)
- Replace duplicate getAccessPackageColor with existing getApColor from utils/colors
- Remove export keyword from internal-only constants in MatrixFilterWizard and MatrixColumnHeaders
- Update all 38+ consumers to import from new locations
