+++ app/ui/src/components/RiskScoringPage.jsx
@@ -1,5 +1,5 @@
 ﻿import { useState, useEffect, useCallback } from 'react';
-import { useAuth } from '../auth/AuthGate';
+import { useAuth } from '../auth/useAuth';
 import { TIER_STYLES } from '../utils/tierStyles';
 
 function TierBadge({ tier }) {
