+++ app/ui/src/components/UserDetailPage.jsx
@@ -1,10 +1,11 @@
 ﻿import { useState, useEffect, useCallback, useMemo } from 'react';
-import { useAuth } from '../auth/AuthGate';
+import { useAuth } from '../auth/useAuth';
 import RiskScoreSection, { RISK_FIELDS } from './RiskScoreSection';
 import { formatDate, computeHistoryDiffs, friendlyLabel } from '../utils/formatters';
 import { CollapsibleSection } from './DetailSection';
 import EntityGraph from './EntityGraph';
-import EntityDetailLayout, { AttributesTable, buildAttributeEntries } from './EntityDetailLayout';
+import EntityDetailLayout, { AttributesTable } from './EntityDetailLayout';
+import { buildAttributeEntries } from '../utils/attributeEntries';
 import ExpandedItemsList from './ExpandedItemsList';
 import RecentChangesSection from './RecentChangesSection';
 import useExpandableGraph from '../hooks/useExpandableGraph';
